Sponsors prep release

+++ src/main/resources/db/migration/V2.9__create_donations_table.sql
@@ -1,7 +1,7 @@
 CREATE TABLE IF NOT EXISTS donations
 (
     id               SERIAL PRIMARY KEY,
-    amount           INTEGER,
+    amount           NUMERIC(18, 2),
     date_of_donation DATE,
     sponsor_id       INTEGER,
     pp_id            TEXT
